Rename to SelectTextFilterOptions
