+++ src/filters/SelectTextFilterOptions/SelectTextFilterOptions.stories.mdx
@@ -2,24 +2,24 @@ import { action } from "@storybook/addon-actions";
 import { Meta, Props, Story } from "@storybook/addon-docs/blocks";
 import { text, withKnobs } from "@storybook/addon-knobs";
 
-import SelectFilterOptions from "./SelectFilterOptions";
+import SelectTextFilterOptions from "./SelectTextFilterOptions";
 
-<Meta title="Filters|SelectFilterOptions" component={SelectFilterOptions} />
+<Meta title="Filters|SelectTextFilterOptions" component={SelectTextFilterOptions} />
 
-# Select Filter Options
+# Select Text Filter Options
 
 ### Display list of checkboxes to select different filter options
 
-The width of SelectFilterOptions takes on the width of its container.
+The width of SelectTextFilterOptions takes on the width of its container.
 
-<Props of={SelectFilterOptions} />
+<Props of={SelectTextFilterOptions} />
 
 ### The number of filter options is > 5
 
 #### Display a input search field to allow users to narrow the list of filter options
 
 <Story name="Enabled Option Query">
-  <SelectFilterOptions
+  <SelectTextFilterOptions
     name={"Committee"}
     state={{}}
     update={action("update-filter-options-state")}
@@ -63,7 +63,7 @@ The width of SelectFilterOptions takes on the width of its container.
 #### Unmatched filter options are disabled
 
 <Story name="Disabled unmatched filter options">
-  <SelectFilterOptions
+  <SelectTextFilterOptions
     name={"Committee"}
     state={{}}
     update={action("update-filter-options-state")}
@@ -109,7 +109,7 @@ The width of SelectFilterOptions takes on the width of its container.
 #### Remove the input search field.
 
 <Story name="Disabled Option Query">
-  <SelectFilterOptions
+  <SelectTextFilterOptions
     name={"Committee"}
     state={{}}
     update={action("update-filter-options-state")}
@@ -147,7 +147,7 @@ The width of SelectFilterOptions takes on the width of its container.
 Change the value of the search string state to change what filter options are disabled.
 
 <Story name="Set Option Query Playground" parameters={{ decorators: [withKnobs] }}>
-  <SelectFilterOptions
+  <SelectTextFilterOptions
     name={"Committee"}
     state={{}}
     update={action("update-filter-options-state")}
