+++ src/filters/SelectTextFilterOptions/getCheckboxText.ts
