+++ src/filters/SelectTextFilterOptions/getSelectedOptions.ts
