+++ src/filters/SelectTextFilterOptions/index.ts
@@ -1,3 +1,3 @@
 export { default as getCheckboxText } from "./getCheckboxText";
 export { default as getSelectedOptions } from "./getSelectedOptions";
-export { default as SelectFilterOptions } from "./SelectFilterOptions";
+export { default as SelectTextFilterOptions } from "./SelectTextFilterOptions";
